Revert "fix coverity issue CID:399957"
This reverts commit de4a0047c5e43fc24b593edf229d83806b64e4b9.

+++ include/libs/nodes/querynodes.h
@@ -38,10 +38,10 @@ typedef struct SRawExprNode {
 } SRawExprNode;
 
 typedef struct SDataType {
-  uint16_t type;
-  uint8_t  precision;
-  uint8_t  scale;
-  int32_t  bytes;
+  uint8_t type;
+  uint8_t precision;
+  uint8_t scale;
+  int32_t bytes;
 } SDataType;
 
 typedef struct SExprNode {
